lib/agents/custom-agent-builder.ts
import { BaseAgent, AgentConfig } from './base';

export interface CustomAgentBuilderConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class CustomAgentBuilderAgent extends BaseAgent {
  private config: CustomAgentBuilderConfig;

  constructor(config: CustomAgentBuilderConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Custom Agent Builder implementation
    await new Promise(resolve => setTimeout(resolve, 1000));
    
    return `Custom Agent Builder completed successfully for: ${input}`;
  }

  async process(data: any): Promise<any> {
    // Advanced custom agent builder processing
    return { status: 'success', data };
  }
}

// FEATURE UPDATE: Custom Agent Builder
// Date: 2024-02-06
import { BaseAgent, AgentConfig } from './base';

export interface CustomAgentBuilderConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class CustomAgentBuilderAgent extends BaseAgent {
  private config: CustomAgentBuilderConfig;

  constructor(config: CustomAgentBuilderConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Custom Agent Builder implementation
    await new Promise...


// FEATURE UPDATE: Custom Agent Builder
// Date: 2024-04-26
import { BaseAgent, AgentConfig } from './base';

export interface CustomAgentBuilderConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class CustomAgentBuilderAgent extends BaseAgent {
  private config: CustomAgentBuilderConfig;

  constructor(config: CustomAgentBuilderConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Custom Agent Builder implementation
    await new Promise...


// FEATURE UPDATE: Custom Agent Builder
// Date: 2024-08-04
import { BaseAgent, AgentConfig } from './base';

export interface CustomAgentBuilderConfig extends AgentConfig {
  mode: 'basic' | 'advanced' | 'enterprise';
  features: string[];
}

export class CustomAgentBuilderAgent extends BaseAgent {
  private config: CustomAgentBuilderConfig;

  constructor(config: CustomAgentBuilderConfig) {
    super(config);
    this.config = config;
  }

  async execute(input: string): Promise<string> {
    // Custom Agent Builder implementation
    await new Promise...
