__tests__/lib/agents/custom-agent-builder.test.ts
import { CustomAgentBuilderAgent } from '@/lib/agents/custom-agent-builder';

describe('Custom Agent BuilderAgent', () => {
  let agent: CustomAgentBuilderAgent;

  beforeEach(() => {
    agent = new CustomAgentBuilderAgent({
      name: 'Test Custom Agent Builder Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      const result = await agent.execute('test input');
      expect(result).toContain('Custom Agent Builder completed successfully');
    });

    it('should handle empty input', async () => {
      const result = await agent.execute('');
      expect(result).toBeDefined();
    });
  });

  describe('process', () => {
    it('should process data correctly', async () => {
      const result = await agent.process({ test: 'data' });
      expect(result.status).toBe('success');
    });
  });
});

// FEATURE UPDATE: Custom Agent Builder
// Date: 2024-02-06
import { CustomAgentBuilderAgent } from '@/lib/agents/custom-agent-builder';

describe('Custom Agent BuilderAgent', () => {
  let agent: CustomAgentBuilderAgent;

  beforeEach(() => {
    agent = new CustomAgentBuilderAgent({
      name: 'Test Custom Agent Builder Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      con...


// FEATURE UPDATE: Custom Agent Builder
// Date: 2024-04-26
import { CustomAgentBuilderAgent } from '@/lib/agents/custom-agent-builder';

describe('Custom Agent BuilderAgent', () => {
  let agent: CustomAgentBuilderAgent;

  beforeEach(() => {
    agent = new CustomAgentBuilderAgent({
      name: 'Test Custom Agent Builder Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      con...


// FEATURE UPDATE: Custom Agent Builder
// Date: 2024-08-04
import { CustomAgentBuilderAgent } from '@/lib/agents/custom-agent-builder';

describe('Custom Agent BuilderAgent', () => {
  let agent: CustomAgentBuilderAgent;

  beforeEach(() => {
    agent = new CustomAgentBuilderAgent({
      name: 'Test Custom Agent Builder Agent',
      model: 'gpt-4',
      temperature: 0.3,
      maxTokens: 2000,
      mode: 'basic',
      features: ['core']
    });
  });

  describe('execute', () => {
    it('should process input successfully', async () => {
      con...
